refactor(app-file-ops): extract rename and bulk-rename workflows
